requirements.txt
psycopg[binary]
pgvector
python-dotenv
sentence-transformers
torch
numpy
requests
beautifulsoup4
lxml
rich